+++ CHANGELOG.md
@@ -1,5 +1,65 @@
 # CHANGELOG
 
+## v2.5.0 (-.-.2021)
+
+## 💔 BC:
+
+-   🔨 All drivers now uses new `config` key to handle local configuration. Involved plugins and drivers:
+-   `plugins`: broadcast, kv
+-   `drivers`: memory, redis, memcached, boltdb.
+
+## ATTENTION!!!, this is configuration BC release, please, update your configuration:
+
+### Old style:
+
+```yaml
+broadcast:
+    default:
+        driver: memory
+        interval: 1
+```
+
+### New style:
+
+```yaml
+broadcast:
+    default:
+        driver: memory
+        config: {} <--------------- NEW
+```
+
+```yaml
+kv:
+    memory-rr:
+        driver: memory
+        config: <--------------- NEW
+            interval: 1
+
+kv:
+    memcached-rr:
+        driver: memcached
+        config: <--------------- NEW
+            addr:
+                - "127.0.0.1:11211"
+```
+
+## 👀 New:
+
+-   ✏️
+
+## 🩹 Fixes:
+
+-   🐛 Fix: local and global configuration parsing
+
+## 📦 Packages:
+
+-   📦
+
+## 📈 Summary:
+
+-   RR Milestone [2.5.0]()
+-   RR-Binary Milestone [2.5.0]()
+
 ## v2.4.1 (13.09.2021)
 
 ## 🩹 Fixes: